create objectToCsv.ts
