+++ packages/trpc/server/routers/viewer/insights/handlers.ts
@@ -7,6 +7,7 @@ import {
   extractDateRangeFromColumnFilters,
   replaceDateRangeColumnFilter,
 } from "@calcom/features/insights/lib/bookingUtils";
+import { objectToCsv } from "@calcom/features/insights/lib/objectToCsv";
 import {
   getTimeView,
   getDateRanges,
@@ -109,28 +110,6 @@ function createInsightsRoutingService(
       columnFilters: input.columnFilters,
     },
   });
-}
-
-function objectToCsv(data: Record<string, string>[]) {
-  if (!data.length) return "";
-
-  const headers = Object.keys(data[0]);
-  const csvRows = [
-    headers.join(","),
-    ...data.map((row) =>
-      headers
-        .map((header) => {
-          const value = row[header]?.toString() || "";
-          // Escape quotes and wrap in quotes if contains comma or newline
-          return value.includes(",") || value.includes("\n") || value.includes('"')
-            ? `"${value.replace(/"/g, '""')}"` // escape double quotes
-            : value;
-        })
-        .join(",")
-    ),
-  ];
-
-  return csvRows.join("\n");
 }
 
 // Handler implementations
